requirements.txt
langchain
langchain-core
langgraph
langchain-openai
streamlit
python-dotenv